requirement.txt
asgiref==3.6.0
Django==4.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
Pillow==9.4.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2022.7
pip install -r requirement.txt